Experiment - prevent having several file handlers opened at the same time
Keeping the file handlers as a member was a good strategy to increase
performance by reducing the number of openings etc.

However, if we are running a project with several experiments and trials,
keeping them open is a bad idea as the system cannot be able to open more
than N files at the same time. Thus, lets open & write & close.