Nodes and Edges are smart pointers now
With the possibility of running an experiment in multiple threads,
the task of removing edges and nodes on-the-fly are tricky.

In order to allow making changes in the graph (nodes and edges)
through a model plugin or GUI, edges and nodes are now shared
pointers and are stored in an unordered_map rather than a vector.

Of course, this change reduces Evoplex's performance.