initial implementation of a minimap for the graph widgets
it places a small minimap in the top left corner of the graph widgets
and is intended to show which part of the graph is being shown in the
screen.

this solution is just experimental and should not be used yet

main issues to investigate:
- what should we do for path graphs? (eg., huge width and small height)
- should we display a screenshot of the whole graph in the map? if so,
how often it should be refreshed?